[Bronze V] Title: 이상한 기호, Time: 124 ms, Memory: 14220 KB, Score: 100 point -BaekjoonHub

+++ 백준/Bronze/15964. 이상한 기호/이상한 기호.java
@@ -8,8 +8,8 @@ public class Main {
         BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
         StringTokenizer token = new StringTokenizer(br.readLine());
 
-        int A = Integer.parseInt(token.nextToken());
-        int B = Integer.parseInt(token.nextToken());
+        Long A = Long.parseLong(token.nextToken());
+        Long B = Long.parseLong(token.nextToken());
 
         System.out.println((A+B)*(A-B));
     }
